+++ game_frontend/src/redux/features/Editor/reducers.js
@@ -3,35 +3,35 @@ import types from './types'
 import { gameTypes } from 'features/Game'
 import { avatarWorkerTypes } from 'features/AvatarWorker'
 import { RunCodeButtonStatus } from 'components/RunCodeButton'
-import { DEFAULT_CODE } from '../constants'
 
 const codeReducer = (state = {}, action) => {
   switch (action.type) {
     case types.GET_CODE_SUCCESS:
       return {
         ...state,
-        codeOnServer: action.payload.code
+        codeOnServer: action.payload.code,
+        starterCode: action.payload.starterCode,
       }
     case types.POST_CODE_REQUEST:
       return {
         ...state,
-        codeToBeSaved: action.payload.code
+        codeToBeSaved: action.payload.code,
       }
     case types.POST_CODE_SUCCESS:
       return {
         ...state,
         codeOnServer: state.codeToBeSaved,
-        codeToBeSaved: null
+        codeToBeSaved: null,
       }
     case types.RESET_CODE:
       return {
         ...state,
-        resetCodeTo: DEFAULT_CODE
+        resetCodeTo: state.starterCode,
       }
     case types.CODE_RESET:
       return {
         ...state,
-        resetCodeTo: null
+        resetCodeTo: null,
       }
     default:
       return state
@@ -43,17 +43,17 @@ const runCodeButtonReducer = (state = {}, action) => {
     case types.POST_CODE_REQUEST:
       return {
         ...state,
-        status: RunCodeButtonStatus.updating
+        status: RunCodeButtonStatus.updating,
       }
     case avatarWorkerTypes.AVATAR_CODE_UPDATED:
       return {
         ...state,
-        status: RunCodeButtonStatus.done
+        status: RunCodeButtonStatus.done,
       }
     case gameTypes.SOCKET_GAME_STATE_RECEIVED:
       return {
         ...state,
-        status: RunCodeButtonStatus.normal
+        status: RunCodeButtonStatus.normal,
       }
     default:
       return state
@@ -62,5 +62,5 @@ const runCodeButtonReducer = (state = {}, action) => {
 
 export default combineReducers({
   code: codeReducer,
-  runCodeButton: runCodeButtonReducer
+  runCodeButton: runCodeButtonReducer,
 })